pharmacies/requirements.txt
asgiref==3.2.3
astroid==2.3.3
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
defusedxml==0.6.0
Django==3.0.1
django-allauth==0.41.0
django-autoslug==1.9.6
django-bootstrap4==1.1.1
django-extensions==2.2.6
django-forms-bootstrap==3.1.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-tastypie==0.14.3
djangorestframework==3.11.0
drf-yasg==1.17.0
idna==2.8
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
lxml==4.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.1
Pillow==7.0.0
pyan==0.0.6
pyan3==1.0.4
pydotplus==2.0.2
Pygments==2.5.2
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pyparsing==2.4.6
python-dateutil==2.8.1
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.6
selenium==3.141.0
simplejson==3.17.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
wrapt==1.11.2
